Optionally call callbacks with pythonic **kwargs
AppDaemon 4.3.0 introduced the option for apps to use true variable
keyword arguments (aka kwargs) in their callback functions. These
functions no longer take the keyword argument dictionary as a positional
parameter, but instead use the '**' operator. This requires the caller
to also use '**' to unpack the dictionary.

I originally tried to implement this to work automatically, using
introspection to check the callback signature. However, I couldn't get
it to work when the callback is a Mock, which seems like it could be a
breaking change.